requirements.txt
pyyaml
requests
termcolor